refactor(apiConnection ping): timer wasn't actually working as expected
When we load the app, we want to make sure the backend is loaded before attempting to show the user any data.

When the app component mounts, we set an interval to ping the backend. When the apiConnection === 1, clear the interval. If we have called on the backend 20 times with no positive response, set the apiConnection to -1 and clear the interval.

Also, the pingApi action should only ever set the `apiConnection` state to true. To set it to false, we need to explicitly call `setApiConnection`