refactor(Toast): move toast into its own component
Thinking about a further refactor: ToastList & toast section in the `state.ui`

`state.ui.toast` would be an array of toasts. toasts contain text and type. When we want to add a toast, we call to the statetree to add it. Success toasts would timeout, loading toasts would disappear when finished loading, and error toasts would need to be x-ed out before disappearing.

This way we can have a little stack of messages/alerts for the whole app.