refactor(selections): add `isLinked` to selections
When you set a working dataset, you set the `peername`, `name`, and now also `isLinked`. We now also save `isLinked` to localStorage.

Currently, the backend api needs to know if `fsi=true` in order to send back the proper response. If the dataset that we have selected is not in the initial `myDatasets` paginated response, we won't know if the dataset is linked or not.

This commit stores that information with the rest of the information that we reference when we want to make an api call.