refactor(fetchWorkingDatasetDetails): move history fetch
this is a bit of a hack, but since, right now, the `fetchWorkingHistory` is the only endpoint that has an error that we are expecting (if the dataset is on the fsi but hasn't been commited yet), let's move that one last so we don't get a longer time from the error being thrown and caught before fetching the next thing.

Also, we probably don't need to `whenOk` these commands, after the first one. They don't need to happen progressively, `fetchWorkingStatus`, `fetchBody`, `fetchWorkingHistory` don't rely on each other.