fix(Dataset, DatasetSidebar): only show status or history if it makes sense
When a dataset is not linked to the fsi, status doesn't mean anything.

When a dataset has no history (is linked to fsi, but user has not saved a commit yet), then history doesn't mean anything.

Only show the history or status tab in the correct case. And, if we switch from one dataset to another, and we are not on a tab that doesn't make sense, switch the tab!