refactor(fetchWorkingDataset): handle 422 'dataset is not linked' error
When we are trying to fetchWorkingDataset, we are gathering information for the dataset status. If the dataset is NOT fsi linked, we don't need any other information besides the peername and dataset name to get all the information we need to populate the history page.

So, let's always assume that when we are trying to `fetchWorkingDataset` we believe that `fsi=true`. We've added the 422 error code on the backend to let us know that the dataset is not linked. (which is different from 404 does not exist, since the dataset does exist in qri, it just isn't linked to the file system)

This commit makes all calls using the `fetchWorkingDataset` action to assume `fsi=true`

When we `DATASET_REQUEST`, we take the peername and name from the action and add it to the state tree

If the `DATASET_FAIL` action emits from this dispatch, but the action errored with a 422 status code, do no clear the full working dataset, allow the peername and name to remain, and ensure that we are marking `fsiPath=''`

In the `fetchWorkingDatasetDetails`, we need to catch the emitted error from `fetchMyDataset`, if it's the 422 error we still need to `fetchWorkingHistory`