fix(api actions): chainSuccess functions need `type: 'SUCCESS'` actions to work
Our chainSuccess functions need all the functions in the chain to emit success actions to work.

However, to cut down on loading in pagination, we check to see if we really should fetch before fetching. The promises that we return from this should emit "success" actions.

We can refactor later when we have a better idea of how this should work.