refactor(ComponentList): allow meta to be clicked even if no meta exists
made some decisions about when and how we show meta is available. This should really be put for further discussion but I think it's fine for now.

when we are in status, meta is always available to click (and is dark to indicate it is 'clickable' MIGHT BE BETTER TO LEAVE IT GREY CAUSE IT'S CONFUSING)