refactor(HistoryItem): add `HistoryItem` model & adjust `mapHistory`
The backend api for returning a the list of a dataset's versions has changed. We are now receving a list of log information rather then a liust of dataset references for each version.

To get the desktop to play nicely, we add a `HistoryItem` interface that will represent how we store each piece of history log information. We also adjust the `mapHistory` function to map each item from the api to the `HistoryItem`

+1 for using mapping functions and having the desktop store data in a way that is useful for the desktop, rather than just storing exactly what is returned in the api. This has made the adjustment very very easy!