feat(FormatConfig): add `FormatConfigFSI` and `FormatConfigHistory`
Because there is no overlap between how we will be displaying the FormatConfig when we are in the status tab and the history tab, to simplify the logic in our components, we are adding `FormatConfigFSI` and `FormatConfigHistory` components. Each imports the `FormatConfigOptions` object from `Structure`. The `FSI` version shows inputs and allows an `onChange` funciton. The `History` version just displays the FormatConfig text.

Another major difference between these two components is the `FSI` version will always show all the options that can exist for this format in the `FormatConfig`, history will only show the options that are set.