refactor(Body): for `status` base format on file extension
When we are in `status`, aka `FSI`, we should base our formatting on the file extension. A user should be able to change the bodyfile and desktop should show the correct formatting options to the user. When we are in `history`, because we know we will always have a structure if we are in history, we should rely on `structure.format` to give us the correct format.