fix(bodyValue): adjust `bodyValue` func for case where we reset pagination
When we are purposely reseting pagination, we have `pageInfo.page === 1`. More subtly, we need to account for when the `previous` body is actually the call `pageInfo.page === 2`, in that case, we shouldn't reset the entire body, but continue through the code path to return the `current` body before the previous body.

Also, let's remove all functions (`concat`, `splice`) that mutate the passed in body, as this may mutate state,