feat(ReadmeHistory): create component that shows versioned `Readme`s
As our first step, let's just have a component that talks right to the `/render/` endpoint and injects the resulting html into the page.

Later, we can add in viewing the raw underlying file. We will need to go through this process when we add transform scripts into the Desktop, but we can punt on this until then.