fix(App, Dataset): event listener leak
The `ipcRenderer.on()` functions that we use to hook into the electron menu are event listeners. We haven't been handling the removal of these listeners in the `App` and `Dataset` components. For each listener we add in `componentDidMount` we need to remove it in `componentDidUnmount`.

This commit also fixes some styling on the Dataset component when a dataset is not `checkedout`