fix(App): fix fetching issues
If there was no peername or name selected, and then you added a dataset, the app would not fetch the newly created dataset.

This fixes it by checking if there is a peername or name set before ignoring the fetch. I there is no peername or name set, we need to fetch the working dataset details.