feat(logging): Log to a file, menu item to export it
When launching the qri backend, always log its output to a temporary file. If a developer wants to see output in a terminal, they can either run `qri connect` themselves, or they can `tail -f` the log. The log filename uses the current day in the filename.

In development mode, try to locate an installed `qri` binary, which makes it easier to develop both desktop and the backend at the same time.

Add a menu item under 'Help' which lets the user copy the debug log to a path.

Output a few useful pieces of info: which version of the qri backend was launched, the path to the the backend, the path to the log file.