fix(commit, MetadataEditor): fix meta editing and the flow after you commit
Adjust the `saveWorkingDatasetAndFetch` and `Commit` components, along with the `mutations` reducer and actions to fix the flow after you commit.

We need to wait for both the commit to be successful and for the history list to update, inorder to show the user a clean commit experience. To do that, we've added a `saveComplete`/`SAVE_COMPLETE` action to the `mutations` reducer that gets called when save & history fetching is complete

Also adjusts the MetadataEditor to fix a content jumping problem