refactor(workbench): construct a location string to shim selection changes
this is a stop-gap while we work on transitioning the workbench away from knowing about selections. Parts of this, however, are the real deal. Being able to quickly check for location-string equality reduces a number of issues in the workbench component update cycle. Let's do it like this from now on