fix(Network): do not show 'clone' button when dataset is in user's collection
This is a bit of a shim. To determine if we should show the 'clone' button, we compare the username and name of the dataset to our list of datasets in our collection. Since we only get back 100 at a time, if a user has more then 100 datasets, we may not get the correct information.

This is fine for now, but should be updated once we settle on how/if we are caching dataset previews in the frontend or if we refactor to fetch at a different point in in the component chain.