fix(network): previews of datasets with JSON bodies now load
issue here was making adjustments to body data structure (adding row numbers) was happening regardless of weather the body is JSON or CSV, which would crash some JSON data structures. The better solution is to decide weather a body is JSON or CSV as early as possible, delegating all details to format-specific components.