refactor(DatasetComponent): now takes a `Dataset` rather then an `CommitDetail`
Refactors `DatasetComponent` to take a `Dataset`, which also means adding `bodyPageInfo`, `peername`, `name`,

The `WorkbenchContainer` now passes down the a status constructed from either the `workingDataset.status` (when the dataset is checked out), or from the `mutations.status.value`, when mutations have already occured, or builds one up from the components of `workingDataset` if the dataset is not checked out, but no mutations have occured yet.

Known bugs: If a dataset starts with no meta, for example, and then you add a title to the meta, but then remove the title, it will still show that you have added a meta. Need a way to detect that the component is empty & to then remove it from the mutations.dataset tree.