fix(HistoryListItem): add `allowDisable` flag to finetune interactions
In some appearances of the `HistoryList` (aka `LogList`), we want to have the option to disable certain items in the list if the version of the dataset is `foreign`. In other appearances, we want the item to still remain clickable and active looking, regardless if the version is `foreign` or not.

When `allowDisable` is true, we can have a `foreign` dataset version, but still take actions on the item