test(e2e): refactor meta editing test so it is reusable
created `commitEdit` that takes a `title` and `message` and commits a version. It then checks that the `title` is the correct title
created `editMeta` that takes a `dataset`, which expects `meta` and `commit` to exist. It enters each existing meta field and tests that they also exist in the history tab once the version is commited