refactor(writeDataset): add `peername` and `name` as params
Rather then pulling peername and name from the state tree inside the action, we are refactoring back to passing peername and name as params.

A later refactor will remove pulling the peername and name from the state tree and instead will be pulled from the `qriRef` param that is passed in as props.