refactor(qriRef): all child dataset components get passed a qriRef prop
We are trying to set the stage for replacing our reliance on the `selections` reducer with relying fully on the router to navigate.

To that effect, we create a `qriRef` object from router props at the `WorkbenchContainer` level. The router props are passed in to the `WorkbenchContainer` straight from the router.

The `Workbench` then passes the `qriRef` to its child components as needed.

We now also have this function `hackAddSelectionsDetailsToQriRef` that pulls from `selections` pieces of data that we will eventually get straight from the router. This is the first thing to resolve when we refactor to relying more on the router.