refactor(backend.js): emit messages to renderer process during backend launch
We refactor the `BackendProcess` class to have specific methods that the main process can use to orchestrate launching the backend app
Because we can only emit events to the renderer process in the main process (so no way to emit them in `BackendProcess`, unless we pass the `mainWindow` down to the `BackendProcess` class) this exposes a bunch of launching functionality to the main process (it was previously all hidden in the `backendProcess.maybeStartup()` function)

In future refactors, am definitely open to pushing `mainWindow` down to `BackendProcess` and dealing with all aspects down there. However, as a first iteration I think this is fine.

The logic for this the start up goes like this:
- if a backend process is already running, error. Catch the error & send a message to the renderer that a qri backend is already running
- check if we have a compatible backend version. If it's too low, send a message to the renderer that the version is incompatible
- check if we need a migration - if we do, send a message to the renderer that we are migrating
- launch the backend