refactor(`main`): need to wait for app to be loaded before coordinating
Both the renderer process and main process need to be on before we can successfully coordinate behavior.

`App` component now emits an `app-fully-loaded` event.

Moved the backend launching process into the ipcMain `app-fully-loaded` event handler. This also means that whenever the app re-loads, we check for a backend connection, and if we don't find one we launch again. Which I think is a nice pick up!

Downside is we get slightly longer loading time in dev mode (since it takes a bit to compile the app), but no discernible changes in prod mode!