refactor(DynamicEditField): conserve typical paste features when sanitizing paste input
We get the range that the user has selected (aka where the selection starts and ends, or, if there is no selection but just a caret, we get the location of the caret), after sanitizing the pasted text (striping all html and new lines), insert the pasted text over/into the selection.

Afterwards, adjust the selection to now select the pasted text.