feat(NewDataset): attach body file to new dataset
adjust a number of aspects related to dataset creation:

* dragging a file over the collection view opens the NewDataset modal
  with the file picker populated
* NewDataset modal now requires a file picker, sets title based
  on inferred filename
* update api action to send a JSON dataset document instead of a
  qri param
* populate meta.title with title from NewDataset modal

A few behaviours need work. It'd be nice if the meta.title that was inferred removed underscores & whatnot. We've also lost import progress tracking, which I think should be re-added as transfer tracking events sent over websockets, but the backend currently doesn't do save progress, so this will have to wait