feat(export): wire up export modal with csv & zip options
we don't have support for per-component export in the backend at the moment,
so I've moved the dataset component picking stuff into it's own react component
for future use in an "advanced" export section.

I'd also like the option to export the entire dataset as a JSON file. I use that
format a bunch, it'd be great to have at some point in the future.