feat(RenameDataset): add modal to rename dataset
Since we already had a pretty slick way to rename a dataset coded up in `DatasetReference` , I decided to just generalize/add functionality that would allow us to use the `DatasetReferenceComponent` as a part of the `RenameModal`. This involved adding optional functions that could inject additional functionality at key points: when the dataset was marked valid/invalid, when the user made any change to the text. The `onSubmit` function became optional, if you add an `onSubmit` function it will submit when the user hits `enter`, otherwise, you control the submitting outside of the component, which is exactly what we do now in the `RenameDataset` modal: we use the `onChange` function to keep track of what the user is typing, and the `isValid` component to keep track of any errors, displaying them to the user.

We've also added the `RenameDataset` modal to storybook.