fix(DatasetLayout): bug that was overriding the qriRef
The DatasetLayout should be getting the qriRef from the `Route` not from the `Dataset` state tree. First, that state tree is only populated if the dataset has a history.

This was causing a bug that was not populating the dataset username or name in the Navbar when a dataset had no history (was fsilinked but no commits).