feat(bulkActionExecuting): add param to state tree if a bulk action is executing
First, a bulk action executing may affect more than just the DatasetList in the future, it potentially adds a strain to the whole app and we may need to check if a bulk action is executing beyond the DatasetList.

Second, the way the remove modal is wired up currently violates our philosophy that all components should 'be responsible for themselves', and it is the only modal that needs outside actions to be passed in.

Third, we now have 2 diffferent ways to remove datasets, but both actions do essentially the same thing, with the same ui required before and after the action: toasts before and after to indicate that the action has happened, and a refresh to the datasets list to reflect the changes.

This commit also removes the `removeDatasetAndFetch` action in favor of the `removeDatasetsAndFetch`. It bypasses the `DatasetList`'s `BulkDatasetActions` function, and ensures all the proper toasts and ui is set. This also means if anything changes with remove datasets, we only have one place to refactor.

A future refactor should extend this to the pull bulk action.