fix(Collection): selected table state
`ReactDataTable` takes a `clearSelectedRows` param. When this value is changed, it triggers the table to clear its internal state of selected rows.

If we remove any datasets, we should clear the selected state. This will also happen when we add datasets, but when we add a dataset we send the user to the dataset page anyway, which would clear the selection regardless.