refactor(ContextMenu): isolate electron elements into platform specific files
Currently, there is no practical use for the context menus in a webapp version. All the actions that occur in context menus only make sense in context of the electron app where we are sure that the user is working locally with a repo that is theirs.

In the future as we extend the ability of the webapp (add login, add tokens etc) so we have a better understanding of who is looking and when, we can put the time into bringing the context menus into the webapp and look for a solution that will satisfy all the use cases.