refactor(*): isolate all electron specific code in platform specific files (#666)
* refactor(*): isolate electron imports in platform specific code with existing (but sometimes stand-in) functions for the web version

This first pass knocks out the simpler cases

* refactor(main.development): remove references to status and history tabs

Removes any references in our menus to the `status` and `history` tabs as well as the dataset view vs the collection view, since the dataset view is now folded into the collection view.

* refactor(App): isolate electron code in platform specific file

This assumes that we have no menu options in a webapp view, nor do we have a full answer yet for exporting a debug log (since the backend is not necessarily local when viewing the webapp), nor do we want to overwrite the normal reload keybinding in someone's browser.

* refactor(ContextMenu): isolate electron elements into platform specific files

Currently, there is no practical use for the context menus in a webapp version. All the actions that occur in context menus only make sense in context of the electron app where we are sure that the user is working locally with a repo that is theirs.

In the future as we extend the ability of the webapp (add login, add tokens etc) so we have a better understanding of who is looking and when, we can put the time into bringing the context menus into the webapp and look for a solution that will satisfy all the use cases.

* refactor(Collection): move dataset actions into `Dataset`

We have some holdover from when Collection and Dataset were seperate views. Moved some listeners from Collection to Dataset, where it makes more sense for them to live.

* fix(LinkDataset): fixes a bug so that when the user choses a new directory to checkout a dataset, the input also shows that directory

* fix(Dataset): adjust effects so keyboard shortcuts work as expected

* refactor(NavBar): use `ExternalLink` component to simplify discord link

Rather than having NavbarItem need to distinguish between a webapp or electron version, let's wrap the item in an `ExternalLink` component and have that component control the linking behavior

* test(e2e): adjust delay timing on csv export