fix(`history`): isolate `history` object & import based on `TARGET_PLATFORM`
Even though we were configuring the store, in the target 'web' version of the build process, to create a `browserRouter` rather than a `hashRouter`, we were still getting hashes in the url. On closer look, it turns out that while we were configuring the store correctly, we were creating a whole separate history to pass to our `ConnectedRouter` in the `App.tsx` file. This history was coming from the `configureStore.development` file, and so was still configured to use `hashHistory`.

Moved the react `history` object into one location, building the hash or browser versions depending on if we are targeting 'web' or 'electron', respectfully.

This allows us to remove the webapp specific store configuration as well as the webapp specific `Index.tsx`, since the history was the only difference in between the `production` and `webapp` versions, and the `webapp/index.tsx` was created specifically to compensate for those differences.