fix(backend): if a port is occupied by a non-qri process, alert the user
If a port is occupied by a non-qri process, or if we cannot tell what process is running that that port, warn the user that they must close down the process at that port for Qri Desktop to work.

If the port is occupied by a qri process and is out of date, kill the process.

Add `PORT` constant to take us one step closer to being able to configure the port on which qri desktop listens.