feat(`StringDiff`): first pass using `react-diff-viewer`
looks pretty good. The font and colors are all very extendable, but I haven't fully experimented with it yet.

Might need to revamp the expand/collapse feature in "Segment", to remove the transformation so that the diff can expand to its full height at all times without having to pass in a `contentHeight`

Also need to figure out how to the get width of the differ to adjust to the width of the page. It's currently too wide, might also just need to change the size and font of the text!