refactor(`Segment`): add `animationOn` flag to `Segment` component
When the `animationOn` flag is true, the Segment has a nice animation sliding in and out. This is limited in that we need to specify a `maxHeight`, which can conflit with our display of line diffing we want for our change report, see so far in the `StringDiff` component.

To allow content to display at its natural height, we have added this `animationOn` flag (that defaults to `true`), so that we can show any diffs without adding any double scrolling to our (eventual) Change Report page