refactor(LabeledStats): adjust to add `delta` value
Add `delta` field to `Stat` interface that is a number intended to show the change from some other value to this current value. Negative values are showing in red, positive in green.

Add `uppercase` flag in the `LabeledStatsProps` interface, that allows us to decide if we want the labels for the stats to be in all uppercase, or just capitalized. Defaults to `true`.