feat(`changes`): add reducers, actions, selectors
Adds the reducers, actions, and selectors for the change report.

Also adjusts the `refStringFromQriRef` & the `qriRefFromString` to work with both the `@` symbol and the text `/at/`. Both of which are expected in different parts of the app (the backend expects `@` when a dataset reference is sent via a body or as params, but `/at/` when it is part of the uri)

Removes the `TempChangeReport` file in favor of the `Changes` component,
that wraps the `ChangeReport` to handle parsing, fetching, and errors