refactor(`DatasetChanges`): adjust route where the `changes` page lives
We now embed the changes into the dataset page itself.

The route is now:
`/collection/changes/:username/:name/at/ipfs/:path`

And we assume that is the "right" side of the change report, looking for the immediately previous version to be the left side.

This designation, that a change report is between one version of a dataset and the previous one, is contained in the wiring surrounding `DatasetChanges`. The `Changes` and `ChangeReport` components are given a `left` and `right`, and we fetch that change report and display the changes. This way, if we want to divorce the notion of `Changes` as being between one version and the immediately previous, we can with little trouble.