refactor(`ChangeReport`): adjust sample api response
The backend api response places the `delta` as a sibling of `left` and `right`.

What was previously named `meta` (as in meta about the change report at a spectific commit) is now named `about`.

We have added another potential `status` type: `missing`. When a component is `missing` it means that it is expected to exist, (typically something computer generated, such as the `Structure` or `Commit`), but cannot be found. In this case, neither the left or right were found & the component should not be displayed at all.
This commit adjusts the `change_report_sample_api_response.json` to match those expectations.